+++ packages/site/src/pages/index.tsx
@@ -188,7 +188,7 @@ const Index = () => {
     console.log("Operation request: ", operationRequest);
     try {
       const provenOperation =
-        await nocturneFrontendSDK!.generateProvenOperation(operationRequest);
+        await nocturneFrontendSDK!.signAndProveOperation(operationRequest);
 
       console.log(provenOperation);
       console.log(